release: 0.5.1 (#4)

+++ include/kb900x_eeprom.h
@@ -72,7 +72,7 @@ typedef struct {
  *
  * \return 0 if no error, else the error code
  */
-int kb900x_eeprom_write(const kb900x_config_t *config, const uint16_t addr, const uint8_t *payload,
+int kb900x_eeprom_write(const kb900x_config_t *config, const uint32_t addr, const uint8_t *payload,
                         const size_t payload_size, const kb900x_eeprom_config_t *eeprom_config);
 
 /** \brief Read data from the EEPROM.
@@ -85,7 +85,7 @@ int kb900x_eeprom_write(const kb900x_config_t *config, const uint16_t addr, cons
  *
  * \return 0 if no error, else the error code
  */
-int kb900x_eeprom_read(const kb900x_config_t *config, const uint16_t addr, const size_t length,
+int kb900x_eeprom_read(const kb900x_config_t *config, const uint32_t addr, const size_t length,
                        uint8_t *result, const kb900x_eeprom_config_t *eeprom_config);
 
 #endif // _KB_EEPROM_H
